added some comments on respective changes you could make to improve this applicatoin
