+++ src/app/models/food.model.ts
@@ -1,0 +1,9 @@
+export class Food {
+  value: number;
+  name: string;
+
+  constructor(value: number, name: string) {
+    this.value = value;
+    this.name = name;
+  }
+}
