+++ src/app/models/foods-result.model.ts
