+++ src/app/models/month.model.ts
@@ -1,0 +1,9 @@
+export class Month {
+  value: string;
+  name: string;
+
+  constructor(value: string, name: string) {
+    this.value = value;
+    this.name = name;
+  }
+}
