--- src/app/search-foods/food.ts
@@ -1,4 +1,0 @@
-export class Food {
-  constructor(public value: number, public name: string) {
-  }
-}
