--- src/app/search-months/month.ts
@@ -1,4 +1,0 @@
-export class Month {
-  constructor(public value: string, public name: string) {
-  }
-}
